refactor: 컴포넌트들 그룹짓기
* analytics: 구글 애널리틱스
* blog: 블로그 포스트 내에 쓰이는 컴포넌트
* card-and-list: 카드와 카드 리스트 컴포넌트
* comments: 댓글 기능
* common: 공통적으로 쓰이는 컴포넌트
* Guestbook: guestbook과 관련된 컴포넌트
* Image: 이미지와 관련된 컴포넌트
* mdxComponents: 마크다운이 컴파일된 html에서 사용할 컴포넌트
* social-icons: 아이콘과 관련된 컴포넌트
* TOC: TOC와 관련된 컴포넌트
