+++ components/blog/ViewCounter.tsx
