+++ components/card-and-list/AboutCard.tsx
