+++ components/common/SectionContainer.tsx
