+++ components/mdxComponents/CustomImage.tsx
@@ -1,11 +1,11 @@
-import AutoImage from '@/components/AutoImage';
+import AutoImage from '@/components/Image/AutoImage';
 
 interface Props {
   src?: string;
   alt?: string;
 }
 
-const Image = ({ src, alt }: Props) => {
+const CustomImage = ({ src, alt }: Props) => {
   return (
     <div className="mb-6 space-y-2 md:space-y-3">
       <AutoImage loading="lazy" src={src as string} alt={alt as string} />
@@ -14,4 +14,4 @@ const Image = ({ src, alt }: Props) => {
   );
 };
 
-export default Image;
+export default CustomImage;
